pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-reviewer-env"
version = "1.0.0"
description = "OpenEnv environment for multi-phase job application review"
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core>=0.2.0",
  "openai>=1.0.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "fastapi>=0.100.0",
  "uvicorn>=0.23.0"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["job_reviewer_env*", "server*"]
